requirements.txt
beautifulsoup4
mitmproxy
lxml
mysql-connector-python
python-dotenv
